+++ src/main/java/com/cg/spblaguna/model/dto/req/RoomReqDTO.java
@@ -1,13 +1,16 @@
 package com.cg.spblaguna.model.dto.req;
 
-import com.cg.spblaguna.model.*;
-import jakarta.validation.constraints.Min;
+import com.cg.spblaguna.model.enumeration.ERoomType;
+import com.cg.spblaguna.model.enumeration.EStatusRoom;
+import com.cg.spblaguna.model.enumeration.EViewType;
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.validator.constraints.Length;
+
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -30,11 +33,11 @@ public class RoomReqDTO {
 
     private Long perTypId;
 
-    private Long pricePerNight;
+    private BigDecimal pricePerNight;
 
-    private Long acreage;
+    private BigDecimal acreage;
 
-    private Long sleep;
+    private Integer sleep;
 
     private String description;
